+++ packages/cli/src/commands/source/push/add.ts
@@ -13,7 +13,7 @@ import {Trackable} from '../../../lib/decorators/preconditions/trackable';
 import {AuthenticatedClient} from '../../../lib/platform/authenticatedClient';
 import {parseAndGetDocumentBuilderFromJSONDocument} from '../../../lib/push/parseFile';
 import {errorMessage, successMessage} from '../../../lib/push/userFeedback';
-import {isDotFile} from '../../../lib/utils/file';
+import {isJsonFile} from '../../../lib/utils/file';
 
 interface AxiosResponse {
   status: number;
@@ -86,7 +86,7 @@ export default class SourcePushAdd extends Command {
 
     const fileNames = flags.folder.flatMap((folder) => {
       const files = readdirSync(folder);
-      return files.map((f) => `${path.join(folder, f)}`);
+      return files.filter(isJsonFile).map((f) => `${path.join(folder, f)}`);
     });
 
     const {send, close} = this.splitByChunkAndUpload(
@@ -99,7 +99,7 @@ export default class SourcePushAdd extends Command {
 
     const folderIterator: string[] = flags.folder.flatMap((folder) =>
       readdirSync(folder)
-        .filter((file) => !isDotFile(file))
+        .filter(isJsonFile)
         .flatMap((file) => path.join(folder, file))
     );
 
@@ -167,7 +167,7 @@ export default class SourcePushAdd extends Command {
     );
   }
 
-  private errorMessageOnAdd(e: Error) {
+  private errorMessageOnAdd(e: unknown) {
     return errorMessage(this, 'Error while trying to add document.', e, {
       exit: true,
     });
@@ -233,8 +233,8 @@ export default class SourcePushAdd extends Command {
         delete: [],
       });
       this.successMessageOnAdd(fileNames, batch.length, res);
-    } catch (e) {
-      this.errorMessageOnAdd(e as Error);
+    } catch (e: unknown) {
+      this.errorMessageOnAdd(e);
     }
   }